requirements.txt
streamlit
moviepy
openai-whisper
google-generativeai
pytube
google-auth
google-cloud-aiplatform
fpdf